SFA: Split code into separate files
