+++ src/StarFoxAdventures/blocks.ts
@@ -1,0 +1,409 @@
+import { GfxDevice, GfxHostAccessPass, GfxTexture, GfxWrapMode, GfxTexFilterMode, GfxMipFilterMode, GfxSampler } from '../gfx/platform/GfxPlatform';
+import { MapInfo } from './maps';
+import { SceneContext } from '../SceneBase';
+import { TextureCollection, ModelInstance, SFARenderer } from './render';
+import { getSubdir } from './resource';
+import { GameInfo } from './scenes';
+import { mat4 } from 'gl-matrix';
+import ArrayBufferSlice from '../ArrayBufferSlice';
+import { GX_VtxDesc, GX_VtxAttrFmt, compileLoadedVertexLayout, compileVtxLoaderMultiVat, LoadedVertexLayout, LoadedVertexData, GX_Array, getAttributeByteSize } from '../gx/gx_displaylist';
+import { hexzero, nArray } from '../util';
+import * as GX from '../gx/gx_enum';
+
+export abstract class BlockFetcher {
+    public abstract getBlock(num: number, trkblk: number, subnum: number): ArrayBufferSlice | null;
+}
+
+interface BlockInfo {
+    base: number;
+    trkblk: number;
+    block: number;
+}
+
+export class BlockCollection {
+    blockRenderers: BlockRenderer[] = [];
+    blockFetcher: BlockFetcher;
+    tex1Tab: ArrayBufferSlice;
+    tex1Bin: ArrayBufferSlice;
+    texColl: TextureCollection;
+
+    public async create(device: GfxDevice, context: SceneContext,  trkblk: number, gameInfo: GameInfo) {
+        const dataFetcher = context.dataFetcher;
+        const pathBase = gameInfo.pathBase;
+        const subdir = getSubdir(trkblk, gameInfo);
+        this.blockFetcher = await gameInfo.makeBlockFetcher(trkblk, dataFetcher, gameInfo);
+        // const tex0Tab = await dataFetcher.fetchData(`${pathBase}/${subdir}/TEX0.tab`);
+        // const tex0Bin = await dataFetcher.fetchData(`${pathBase}/${subdir}/TEX0.bin`);
+        this.tex1Tab = await dataFetcher.fetchData(`${pathBase}/${subdir}/TEX1.tab`);
+        this.tex1Bin = await dataFetcher.fetchData(`${pathBase}/${subdir}/TEX1.bin`);
+        this.texColl = new TextureCollection(this.tex1Tab, this.tex1Bin);
+    }
+
+    public getBlockRenderer(device: GfxDevice, blockNum: number, trkblk: number, subnum: number): BlockRenderer | null {
+        if (this.blockRenderers[blockNum] === undefined) {
+            const uncomp = this.blockFetcher.getBlock(blockNum, trkblk, subnum);
+            if (uncomp === null)
+                return null;
+            this.blockRenderers[blockNum] = new BlockRenderer(device, uncomp, this.texColl);
+        }
+
+        return this.blockRenderers[blockNum];
+    }
+}
+
+export function getBlockInfo(mapsBin: DataView, mapInfo: MapInfo, x: number, y: number, trkblkTab: DataView, locationNum: number): BlockInfo {
+    const blockIndex = y * mapInfo.blockCols + x;
+    const blockInfo = mapsBin.getUint32(mapInfo.blockTableOffset + 4 * blockIndex);
+    const base = (blockInfo >>> 17) & 0x3F;
+    const trkblk = (blockInfo >>> 23);
+    let block;
+    if (trkblk == 0xff) {
+        block = -1;
+    } else {
+        try {
+            block = base + trkblkTab.getUint16(trkblk * 2);
+        } catch (e) {
+            block = -1
+        }
+    }
+    return {base, trkblk, block};
+}
+
+// Reads bitfields by pulling from the low bits of each byte in sequence
+class LowBitReader {
+    dv: DataView
+    offs: number
+    num: number
+    buf: number
+
+    constructor(dv: DataView, offs: number = 0) {
+        this.dv = dv;
+        this.offs = offs;
+        this.num = 0;
+        this.buf = 0;
+    }
+
+    peek(bits: number): number {
+        while (this.num < bits) {
+            this.buf |= this.dv.getUint8(this.offs) << this.num
+            this.offs++;
+            this.num += 8;
+        }
+        return this.buf & ((1<<bits)-1);
+    }
+
+    drop(bits: number) {
+        this.peek(bits); // Ensure buffer has bits to drop
+        this.buf >>>= bits
+        this.num -= bits
+    }
+
+    get(bits: number): number {
+        const x = this.peek(bits)
+        this.drop(bits)
+        return x
+    }
+}
+
+export class BlockRenderer {
+    models: ModelInstance[] = [];
+    yTranslate: number = 0;
+
+    constructor(device: GfxDevice, blockData: ArrayBufferSlice, texColl: TextureCollection) {
+        let offs = 0;
+        const blockDv = blockData.createDataView();
+
+        // FIXME: This field is NOT a model type and doesn't reliably indicate
+        // the type of model.
+        const modelType = blockDv.getUint16(4);
+        let fields;
+        switch (modelType) {
+        case 0:
+            fields = {
+                texOffset: 0x54,
+                texCount: 0xa0,
+                posOffset: 0x58,
+                posCount: 0x90,
+                clrOffset: 0x5c,
+                clrCount: 0x94,
+                texcoordOffset: 0x60,
+                texcoordCount: 0x96,
+                shaderOffset: 0x64,
+                shaderCount: 0xa0, // Polygon attributes and material information
+                shaderSize: 0x40,
+                listOffset: 0x68,
+                listCount: 0x9f,
+                listSize: 0x34,
+                numListBits: 8, // ??? should be 6 according to decompilation of demo????
+                numLayersOffset: 0x3b,
+                bitstreamOffset: 0x74, // Whoa...
+                // FIXME: There are three bitstreams, probably for opaque and transparent objects
+                bitstreamByteCount: 0x84,
+            };
+            break;
+        case 8:
+        case 264:
+            fields = {
+                texOffset: 0x54,
+                texCount: 0xa0,
+                posOffset: 0x58,
+                posCount: 0x90,
+                clrOffset: 0x5c,
+                clrCount: 0x94,
+                texcoordOffset: 0x60,
+                texcoordCount: 0x96,
+                shaderOffset: 0x64,
+                shaderCount: 0xa2,
+                shaderSize: 0x44,
+                listOffset: 0x68,
+                listCount: 0xa1, // TODO
+                listSize: 0x1c,
+                numListBits: 8,
+                numLayersOffset: 0x41,
+                bitstreamOffset: 0x78,
+                bitstreamByteCount: 0x84,
+            };
+            break;
+        default:
+            throw Error(`Model type ${modelType} not implemented`);
+        }
+
+        // @0x8: data size
+        // @0xc: 4x3 matrix (placeholder; always zeroed in files)
+        // @0x8e: y translation (up/down)
+
+        this.yTranslate = blockDv.getInt16(0x8e); // TODO: fix for demo
+
+        //////////// TEXTURE STUFF TODO: move somewhere else
+
+        const texOffset = blockDv.getUint32(fields.texOffset);
+        const texCount = blockDv.getUint8(fields.texCount);
+        // const texCount = 8; // XXX: can't find the damn field...
+        // console.log(`Loading ${texCount} texture infos from 0x${texOffset.toString(16)}`);
+        const texIds: number[] = [];
+        for (let i = 0; i < texCount; i++) {
+            const texIdFromFile = blockDv.getUint32(texOffset + i * 4);
+            texIds.push(texIdFromFile);
+        }
+        // console.log(`tex ids: ${JSON.stringify(texIds)}`);
+
+        //////////////////////////
+
+        const posOffset = blockDv.getUint32(fields.posOffset);
+        const posCount = blockDv.getUint16(fields.posCount);
+        // const posCount = 3200; // XXX: can't find the damn field...
+        // console.log(`Loading ${posCount} positions from 0x${posOffset.toString(16)}`);
+        const vertBuffer = blockData.subarray(posOffset, posCount * 3 * 2);
+
+        const clrOffset = blockDv.getUint32(fields.clrOffset);
+        const clrCount = blockDv.getUint16(fields.clrCount);
+        // const clrCount = 3200;
+        // console.log(`Loading ${clrCount} colors from 0x${clrOffset.toString(16)}`);
+        const clrBuffer = blockData.subarray(clrOffset, clrCount * 2);
+
+        const texcoordOffset = blockDv.getUint32(fields.texcoordOffset);
+        const texcoordCount = blockDv.getUint16(fields.texcoordCount);
+        // const coordCount = 3200;
+        // console.log(`Loading ${coordCount} texcoords from 0x${coordOffset.toString(16)}`);
+        const texcoordBuffer = blockData.subarray(texcoordOffset, texcoordCount * 2 * 2);
+
+        const shaderOffset = blockDv.getUint32(fields.shaderOffset);
+        const shaderCount = blockDv.getUint8(fields.shaderCount);
+        // console.log(`Loading ${polyCount} polytypes from 0x${polyOffset.toString(16)}`);
+
+        interface Shader {
+            numLayers: number;
+            tex0Num: number;
+            tex1Num: number;
+            hasTexCoord: boolean[];
+            enableCull: boolean;
+        }
+
+        const shaders: Shader[] = [];
+        offs = shaderOffset;
+        for (let i = 0; i < shaderCount; i++) {
+            const shader = {
+                numLayers: 0,
+                hasTexCoord: nArray(8, () => false),
+                tex0Num: -1,
+                tex1Num: -1,
+                enableCull: false,
+            };
+            // console.log(`parsing polygon attributes ${i} from 0x${offs.toString(16)}`);
+            shader.tex0Num = blockDv.getUint32(offs + 0x24);
+            //polyType.tex1Num = blockDv.getUint32(offs + 0x2C);
+            shader.tex1Num = blockDv.getUint32(offs + 0x34); // According to decompilation
+            shader.numLayers = blockDv.getUint8(offs + fields.numLayersOffset);
+            for (let j = 0; j < shader.numLayers; j++) {
+                shader.hasTexCoord[j] = true;
+            }
+            // FIXME: find this field's offset for demo files
+            shader.enableCull = (blockDv.getUint32(offs + 0x3c) & 0x8) != 0;
+            // polyType.enableCull = false;
+            
+            // console.log(`PolyType: ${JSON.stringify(polyType)}`);
+            // console.log(`PolyType tex0: ${decodedTextures[polyType.tex0Num]}, tex1: ${decodedTextures[polyType.tex1Num]}`);
+            shaders.push(shader);
+            offs += fields.shaderSize;
+        }
+        
+        const vcd: GX_VtxDesc[] = [];
+        const vat: GX_VtxAttrFmt[][] = nArray(8, () => []);
+        for (let i = 0; i <= GX.Attr.MAX; i++) {
+            vcd[i] = { type: GX.AttrType.NONE };
+            for (let j = 0; j < 8; j++) {
+                vat[j][i] = { compType: GX.CompType.U8, compShift: 0, compCnt: 0 };
+            }
+        }
+
+        // vcd[GX.Attr.PNMTXIDX].type = GX.AttrType.DIRECT;
+        vcd[GX.Attr.POS].type = GX.AttrType.DIRECT;
+        vcd[GX.Attr.CLR0].type = GX.AttrType.DIRECT;
+        vcd[GX.Attr.TEX0].type = GX.AttrType.DIRECT;
+
+        // TODO: Implement normals and lighting
+        vat[0][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 0, compCnt: GX.CompCnt.POS_XYZ };
+        vat[0][GX.Attr.CLR0] = { compType: GX.CompType.RGBA8, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[0][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 7, compCnt: GX.CompCnt.TEX_ST };
+        
+        vat[1][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 2, compCnt: GX.CompCnt.POS_XYZ };
+        vat[1][GX.Attr.CLR0] = { compType: GX.CompType.RGBA8, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[1][GX.Attr.TEX0] = { compType: GX.CompType.F32, compShift: 0, compCnt: GX.CompCnt.TEX_ST };
+        
+        vat[2][GX.Attr.POS] = { compType: GX.CompType.F32, compShift: 0, compCnt: GX.CompCnt.POS_XYZ };
+        vat[2][GX.Attr.CLR0] = { compType: GX.CompType.RGBA8, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[2][GX.Attr.TEX0] = { compType: GX.CompType.F32, compShift: 0, compCnt: GX.CompCnt.TEX_ST };
+        vat[2][GX.Attr.TEX1] = { compType: GX.CompType.F32, compShift: 0, compCnt: GX.CompCnt.TEX_ST };
+
+        vat[3][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.POS_XYZ };
+        vat[3][GX.Attr.CLR0] = { compType: GX.CompType.RGBA4, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[3][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[3][GX.Attr.TEX1] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[3][GX.Attr.TEX2] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[3][GX.Attr.TEX3] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        
+        vat[4][GX.Attr.POS] = { compType: GX.CompType.F32, compShift: 8, compCnt: GX.CompCnt.POS_XYZ };
+        vat[4][GX.Attr.CLR0] = { compType: GX.CompType.RGBA8, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[4][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 7, compCnt: GX.CompCnt.TEX_ST };
+
+        vat[5][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 3, compCnt: GX.CompCnt.POS_XYZ };
+        vat[5][GX.Attr.CLR0] = { compType: GX.CompType.RGBA4, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[5][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.TEX_ST };
+        vat[5][GX.Attr.TEX1] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.TEX_ST };
+        vat[5][GX.Attr.TEX2] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.TEX_ST };
+        vat[5][GX.Attr.TEX3] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.TEX_ST };
+
+        vat[6][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 8, compCnt: GX.CompCnt.POS_XYZ };
+        vat[6][GX.Attr.CLR0] = { compType: GX.CompType.RGBA4, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[6][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[6][GX.Attr.TEX1] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[6][GX.Attr.TEX2] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[6][GX.Attr.TEX3] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+
+        vat[7][GX.Attr.POS] = { compType: GX.CompType.S16, compShift: 0, compCnt: GX.CompCnt.POS_XYZ };
+        vat[7][GX.Attr.CLR0] = { compType: GX.CompType.RGBA4, compShift: 0, compCnt: GX.CompCnt.CLR_RGBA };
+        vat[7][GX.Attr.TEX0] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[7][GX.Attr.TEX1] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[7][GX.Attr.TEX2] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+        vat[7][GX.Attr.TEX3] = { compType: GX.CompType.S16, compShift: 10, compCnt: GX.CompCnt.TEX_ST };
+
+        const listOffset = blockDv.getUint32(fields.listOffset);
+        const listCount = blockDv.getUint8(fields.listCount);
+        // console.log(`Loading ${chunkCount} display lists from 0x${chunkOffset.toString(16)}`);
+
+        const bitstreamOffset = blockDv.getUint32(fields.bitstreamOffset);
+        const bitstreamCount = blockDv.getUint16(fields.bitstreamByteCount);
+        // console.log(`Loading ${bitsCount} bits from 0x${bitsOffset.toString(16)}`);
+
+        const bits = new LowBitReader(blockDv, bitstreamOffset);
+        let done = false;
+        let curShader = 0;
+        while (!done) {
+            const opcode = bits.get(4);
+            switch (opcode) {
+            case 1: // Set polygon type
+                curShader = bits.get(6);
+                // console.log(`setting poly type ${curPolyType}`);
+                break;
+            case 2: // Geometry
+                const listNum = bits.get(fields.numListBits);
+                // console.log(`Drawing display list #${chunkNum}`);
+                if (listNum >= listCount) {
+                    console.warn(`Can't draw display list #${listNum} (out of range)`);
+                    continue;
+                }
+                offs = listOffset + listNum * fields.listSize;
+                const dlOffset = blockDv.getUint32(offs);
+                const dlSize = blockDv.getUint16(offs + 4);
+                // console.log(`DL offset 0x${dlOffset.toString(16)} size 0x${dlSize.toString(16)}`);
+                const displayList = blockData.subarray(dlOffset, dlSize);
+
+                const vtxArrays: GX_Array[] = [];
+                vtxArrays[GX.Attr.POS] = { buffer: vertBuffer, offs: 0, stride: 6 /*getAttributeByteSize(vat[0], GX.Attr.POS)*/ };
+                vtxArrays[GX.Attr.CLR0] = { buffer: clrBuffer, offs: 0, stride: 2 /*getAttributeByteSize(vat[0], GX.Attr.CLR0)*/ };
+                for (let t = 0; t < 8; t++) {
+                    vtxArrays[GX.Attr.TEX0 + t] = { buffer: texcoordBuffer, offs: 0, stride: 4 /*getAttributeByteSize(vat[0], GX.Attr.TEX0)*/ };
+                }
+
+                try {
+                    const shader = shaders[curShader];
+                    const newModel = new ModelInstance(vtxArrays, vcd, vat, displayList, shader.enableCull);
+                    if (shader.numLayers == 2) {
+                        newModel.setTextures([
+                            texColl.getTexture(device, texIds[shader.tex0Num]),
+                            texColl.getTexture(device, texIds[shader.tex1Num]),
+                        ]);
+                    } else if (shader.numLayers == 1) {
+                        newModel.setTextures([
+                            texColl.getTexture(device, texIds[shader.tex0Num]), // ???
+                        ]);
+                    }
+                    this.models.push(newModel);
+                } catch (e) {
+                    console.error(e);
+                }
+                break;
+            case 3: // Set vertex attributes
+                const posDesc = bits.get(1);
+                const colorDesc = bits.get(1);
+                const texCoordDesc = bits.get(1);
+                vcd[GX.Attr.POS].type = posDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                vcd[GX.Attr.NRM].type = GX.AttrType.NONE; // Normal is not used in Star Fox Adventures (?)
+                vcd[GX.Attr.CLR0].type = colorDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                if (shaders[curShader].hasTexCoord[0]) {
+                    // Note: texCoordDesc applies to all texture coordinates in the vertex
+                    for (let t = 0; t < 8; t++) {
+                        if (shaders[curShader].hasTexCoord[t]) {
+                            vcd[GX.Attr.TEX0 + t].type = texCoordDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                        } else {
+                            vcd[GX.Attr.TEX0 + t].type = GX.AttrType.NONE;
+                        }
+                    }
+                }
+                break;
+            case 4: // Set weights (skipped by SFA block renderer)
+                const numWeights = bits.get(4);
+                for (let i = 0; i < numWeights; i++) {
+                    bits.get(8);
+                }
+                break;
+            case 5: // End
+                done = true;
+                break;
+            default:
+                console.warn(`Skipping unknown model bits opcode ${opcode}`);
+                break;
+            }
+        }
+    }
+
+    public addToRenderer(renderer: SFARenderer, modelMatrix: mat4) {
+        for (let i = 0; i < this.models.length; i++) {
+            const trans = mat4.create();
+            mat4.fromTranslation(trans, [0, this.yTranslate, 0]);
+            const matrix = mat4.create();
+            mat4.mul(matrix, modelMatrix, trans);
+            renderer.addModel(this.models[i], matrix);
+        }
+    }
+}
